registry: build manifest index
Create a new `manifest` directory that will serve as the index of
available manifests. For each stored manifest, a directory is
created with the digest of the manifest. In it a `manifest.json`
will link to the actual manifest blob data. In the future, links
to the configuration and actual layer blobs will also be stored
in there. The idea is to use this as a simple form of reference
counting for the blobs in the blob-store.